Add input validation
This adds a function for Requirement 2.x in Homework 04.
Note that it returns "Not A Triangle" for the case where the numbers provided
are "too big". It may be preferable to have a separate error message that
indicates the exact reason for failure, but the requirements did not specify
how the error conditions were to be handled.